CLDR-14590 Modernize Ajax; prepare revision for pending votes
-Move voting-related code from cldrSurvey.js to new cldrVote.js

-Add stub for cldrVote.pending, call it from handleChangedLocaleStamp

-This will compensate for queueCount call removed from handleChangedLocaleStamp in last PR
